Change EstablishmentGroup to Group and return group with results
