--- back/src/domain/offer/entities/EstablishmentGroupEntity.ts
@@ -1,11 +1,0 @@
-import {
-  EstablishmentGroupName,
-  EstablishmentGroupSlug,
-  SiretDto,
-} from "shared";
-
-export type EstablishmentGroupEntity = {
-  slug: EstablishmentGroupSlug;
-  name: EstablishmentGroupName;
-  sirets: SiretDto[];
-};
