+++ back/src/domain/offer/entities/GroupEntity.ts
@@ -1,0 +1,7 @@
+import { GroupName, GroupSlug, SiretDto } from "shared";
+
+export type GroupEntity = {
+  slug: GroupSlug;
+  name: GroupName;
+  sirets: SiretDto[];
+};
