--- back/src/domain/offer/ports/EstablishmentGroupRepository.ts
@@ -1,11 +1,0 @@
-import { EstablishmentGroupSlug, SearchResultDto, SiretDto } from "shared";
-import { EstablishmentGroupEntity } from "../entities/EstablishmentGroupEntity";
-
-export interface EstablishmentGroupRepository {
-  groupsWithSiret(siret: SiretDto): Promise<EstablishmentGroupEntity[]>;
-  findSearchImmersionResultsBySlug(
-    slug: EstablishmentGroupSlug,
-  ): Promise<SearchResultDto[]>;
-
-  save(group: EstablishmentGroupEntity): Promise<void>;
-}
