+++ back/src/domain/offer/ports/GroupRepository.ts
@@ -1,0 +1,8 @@
+import { GroupSlug, SearchResultDto, SiretDto } from "shared";
+import { GroupEntity } from "../entities/GroupEntity";
+
+export interface GroupRepository {
+  groupsWithSiret(siret: SiretDto): Promise<GroupEntity[]>;
+  findSearchResultsBySlug(slug: GroupSlug): Promise<SearchResultDto[]>;
+  save(group: GroupEntity): Promise<void>;
+}
